refactor: move Redis initialization to a separate module
